+++ src/init.c
@@ -3,8 +3,10 @@
 
 #include "keypress.h"
 #include "keypress-internal.h"
+#include "cleancall.h"
 
 static const R_CallMethodDef callMethods[]  = {
+  CLEANCALL_METHOD_RECORD,
   {"keypress", (DL_FUNC) &keypress, 1},
   {"save_term_status", (DL_FUNC) &save_term_status, 0},
   {"restore_term_status", (DL_FUNC) &restore_term_status, 0},
@@ -18,6 +20,7 @@ void R_init_keypress(DllInfo *dll) {
   R_registerRoutines(dll, NULL, callMethods, NULL, NULL);
   R_useDynamicSymbols(dll, FALSE);
   R_forceSymbols(dll, TRUE);
+  cleancall_init();
 
   R_RegisterCCallable("keypress", "keypress", (DL_FUNC) keypress);
   R_RegisterCCallable("keypress_read", "keypress_read", (DL_FUNC) keypress_read);